refactor(sections): extract the flag-list repairs into repair.rs
Pure motion: the repeated-character shape (-v/-vv), the single-dash
long-option shape (-help, -number=N), and the glued-token index both
consult for evidence.
